Move rotation conversions from drakeGeometryUtil into drake/math.
