+++ drake/math/CMakeLists.txt
@@ -1,17 +1,25 @@
 # Source files used to build drakeMath.
 set(sources
   autodiff.cc
+  axis_angle.cc
+  cylindrical.cc
   expmap.cc
   gradient.cc
-  quaternion.cc)
+  quaternion.cc
+  roll_pitch_yaw.cc
+  rotation_matrix.cc)
 
 # Headers that should be installed with Drake so that they
 # are available elsewhere via #include.
 set(installed_headers
   autodiff.h
+  axis_angle.h
+  cylindrical.h
   expmap.h
   gradient.h
-  quaternion.h)
+  quaternion.h
+  roll_pitch_yaw.h
+  rotation_matrix.h)
 
 # Headers that are needed by code here but should not
 # be exposed anywhere else.
