+++ drake/math/axis_angle.cc
@@ -1,0 +1,4 @@
+// For now, this is an empty .cc file that only serves to confirm
+// axis_angle.h is a stand-alone header.
+
+#include "drake/math/axis_angle.h"
