+++ drake/math/quaternion.h
@@ -15,7 +15,6 @@ namespace math {
 template <typename Derived>
 Vector4<typename Derived::Scalar> quatConjugate(
     const Eigen::MatrixBase<Derived>& q) {
-  using namespace Eigen;
   static_assert(Derived::SizeAtCompileTime == 4, "Wrong size.");
   Vector4<typename Derived::Scalar> q_conj;
   q_conj << q(0), -q(1), -q(2), -q(3);
@@ -26,7 +25,6 @@ template <typename Derived1, typename Derived2>
 Vector4<typename Derived1::Scalar> quatProduct(
     const Eigen::MatrixBase<Derived1>& q1,
     const Eigen::MatrixBase<Derived2>& q2) {
-  using namespace Eigen;
   static_assert(Derived1::SizeAtCompileTime == 4, "Wrong size.");
   static_assert(Derived2::SizeAtCompileTime == 4, "Wrong size.");
 
@@ -45,7 +43,6 @@ template <typename DerivedQ, typename DerivedV>
 Vector3<typename DerivedV::Scalar> quatRotateVec(
     const Eigen::MatrixBase<DerivedQ>& q,
     const Eigen::MatrixBase<DerivedV>& v) {
-  using namespace Eigen;
   static_assert(DerivedQ::SizeAtCompileTime == 4, "Wrong size.");
   static_assert(DerivedV::SizeAtCompileTime == 3, "Wrong size.");
 
@@ -78,7 +75,8 @@ typename Derived1::Scalar quatDiffAxisInvar(
 
 template <typename Derived>
 typename Derived::Scalar quatNorm(const Eigen::MatrixBase<Derived>& q) {
-  return std::acos(q(0));
+  using std::acos;
+  return acos(q(0));
 }
 
 /**
@@ -100,6 +98,9 @@ template <typename Derived1, typename Derived2, typename Scalar>
 Vector4<Scalar> Slerp(const Eigen::MatrixBase<Derived1>& q1,
                       const Eigen::MatrixBase<Derived2>& q2,
                       const Scalar& interpolation_parameter) {
+  using std::acos;
+  using std::sin;
+
   // Compute the quaternion inner product
   auto lambda = (q1.transpose() * q2).value();
   int q2_sign;
@@ -120,8 +121,8 @@ Vector4<Scalar> Slerp(const Eigen::MatrixBase<Derived1>& q1,
     r = 1.0 - interpolation_parameter;
     s = interpolation_parameter;
   } else {
-    Scalar alpha = std::acos(lambda);
-    Scalar gamma = 1.0 / std::sin(alpha);
+    Scalar alpha = acos(lambda);
+    Scalar gamma = 1.0 / sin(alpha);
     r = std::sin((1.0 - interpolation_parameter) * alpha) * gamma;
     s = std::sin(interpolation_parameter * alpha) * gamma;
   }
@@ -129,6 +130,79 @@ Vector4<Scalar> Slerp(const Eigen::MatrixBase<Derived1>& q1,
   auto ret = (q1 * r).eval();
   ret += q2 * (q2_sign * s);
   return ret;
+}
+
+template <typename Derived>
+Vector4<typename Derived::Scalar> quat2axis(
+    const Eigen::MatrixBase<Derived>& q) {
+  using std::sqrt;
+  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
+  auto q_normalized = q.normalized();
+  auto s = sqrt(1.0 - q_normalized(0) * q_normalized(0)) +
+           std::numeric_limits<typename Derived::Scalar>::epsilon();
+  Vector4<typename Derived::Scalar> a;
+
+  a << q_normalized.template tail<3>() / s, 2.0 * std::acos(q_normalized(0));
+  return a;
+}
+
+template <typename Derived>
+Matrix3<typename Derived::Scalar> quat2rotmat(
+    const Eigen::MatrixBase<Derived>& q) {
+  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
+  auto q_normalized = q.normalized();
+  auto w = q_normalized(0);
+  auto x = q_normalized(1);
+  auto y = q_normalized(2);
+  auto z = q_normalized(3);
+
+  Matrix3<typename Derived::Scalar> M;
+  M.row(0) << w * w + x * x - y * y - z * z, 2.0 * x * y - 2.0 * w * z,
+      2.0 * x * z + 2.0 * w * y;
+  M.row(1) << 2.0 * x * y + 2.0 * w * z, w * w + y * y - x * x - z * z,
+      2.0 * y * z - 2.0 * w * x;
+  M.row(2) << 2.0 * x * z - 2.0 * w * y, 2.0 * y * z + 2.0 * w * x,
+      w * w + z * z - x * x - y * y;
+
+  return M;
+}
+
+template <typename Derived>
+Vector3<typename Derived::Scalar> quat2rpy(
+    const Eigen::MatrixBase<Derived>& q) {
+  using std::asin;
+  using std::atan2;
+  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
+  auto q_normalized = q.normalized();
+  auto w = q_normalized(0);
+  auto x = q_normalized(1);
+  auto y = q_normalized(2);
+  auto z = q_normalized(3);
+
+  Vector3<typename Derived::Scalar> ret;
+  ret << atan2(2.0 * (w * x + y * z), w * w + z * z - (x * x + y * y)),
+      asin(2.0 * (w * y - z * x)),
+      atan2(2.0 * (w * z + x * y), w * w + x * x - (y * y + z * z));
+  return ret;
+}
+
+template <typename Derived>
+Eigen::Quaternion<typename Derived::Scalar> quat2eigenQuaternion(
+    const Eigen::MatrixBase<Derived>& q) {
+  // The Eigen Quaterniond constructor when used with 4 arguments, uses the (w,
+  // x, y, z) ordering, just as we do.
+  // HOWEVER: when the constructor is called on a 4-element Vector, the elements
+  // must be in (x, y, z, w) order.
+  // So, the following two calls will give you the SAME quaternion:
+  // Quaternion<double>(q(0), q(1), q(2), q(3));
+  // Quaternion<double>(Vector4d(q(3), q(0), q(1), q(2)))
+  // which is gross and will cause you much pain.
+  // see:
+  // http://eigen.tuxfamily.org/dox/classEigen_1_1Quaternion.html#a91b6ea2cac13ab2d33b6e74818ee1490
+  //
+  // This method takes a nice, normal (w, x, y, z) order vector and gives you
+  // the Quaternion you expect.
+  return Eigen::Quaternion<typename Derived::Scalar>(q(0), q(1), q(2), q(3));
 }
 
 }  // namespace math
