+++ drake/math/roll_pitch_yaw.cc
@@ -1,0 +1,4 @@
+// For now, this is an empty .cc file that only serves to confirm
+// roll_pitch_yaw.h is a stand-alone header.
+
+#include "drake/math/roll_pitch_yaw.h"
