+++ drake/math/rotation_matrix.cc
@@ -1,0 +1,4 @@
+// For now, this is an empty .cc file that only serves to confirm
+// rotation_matrix.h is a stand-alone header.
+
+#include "drake/math/rotation_matrix.h"
