+++ drake/systems/plants/rigid_body_frame.cc
@@ -1,10 +1,10 @@
 #include "drake/systems/plants/RigidBodyFrame.h"
 
-#include "drake/util/drakeGeometryUtil.h"
+#include "drake/math/roll_pitch_yaw.h"
 
 RigidBodyFrame::RigidBodyFrame(const std::string& _name, RigidBody* _body,
                                const Eigen::Vector3d& xyz,
                                const Eigen::Vector3d& rpy)
     : name(_name), body(_body) {
-  transform_to_body.matrix() << rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
+  transform_to_body.matrix() << drake::math::rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
 }
