+++ drake/systems/plants/test/rigid_body_tree/rigid_body_tree_test.cc
@@ -3,6 +3,7 @@
 #include <gtest/gtest.h>
 
 #include "drake/common/eigen_types.h"
+#include "drake/math/roll_pitch_yaw.h"
 #include "drake/Path.h"
 #include "drake/systems/plants/parser_urdf.h"
 #include "drake/systems/plants/RigidBodyTree.h"
@@ -65,7 +66,7 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointNoOffset) {
   // Adds floating joints that connect r1b1 and r2b1 to the rigid body tree's
   // world at zero offset.
   tree->AddFloatingJoint(DrakeJoint::QUATERNION,
-                        {r1b1->body_index, r2b1->body_index});
+                         {r1b1->body_index, r2b1->body_index});
 
   // Verfies that the two rigid bodies are located in the correct place.
   const DrakeJoint& jointR1B1 = tree->FindBody("body1", "robot1")->getJoint();
@@ -92,7 +93,8 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWithOffset) {
     Eigen::Vector3d xyz, rpy;
     xyz << 1, 1, 1;
     rpy = Eigen::Vector3d::Zero();
-    T_r1and2_to_world.matrix() << rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
+    T_r1and2_to_world.matrix()
+        << drake::math::rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
   }
 
   auto weld_to_frame = std::allocate_shared<RigidBodyFrame>(
@@ -100,7 +102,7 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWithOffset) {
       T_r1and2_to_world);
 
   tree->AddFloatingJoint(DrakeJoint::QUATERNION,
-                        {r1b1->body_index, r2b1->body_index}, weld_to_frame);
+                         {r1b1->body_index, r2b1->body_index}, weld_to_frame);
 
   // Verfies that the two rigid bodies are located in the correct place.
   const DrakeJoint& jointR1B1 = tree->FindBody("body1", "robot1")->getJoint();
@@ -134,7 +136,7 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
     Eigen::Vector3d xyz, rpy;
     xyz << 1, 1, 1;
     rpy = Eigen::Vector3d::Zero();
-    T_r2_to_r1.matrix() << rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
+    T_r2_to_r1.matrix() << drake::math::rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
   }
 
   auto r2b1_weld = std::allocate_shared<RigidBodyFrame>(
@@ -155,7 +157,8 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
     Eigen::Vector3d xyz, rpy;
     xyz << 2, 2, 2;
     rpy = Eigen::Vector3d::Zero();
-    T_r3_and_r4_to_r2.matrix() << rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
+    T_r3_and_r4_to_r2.matrix()
+        << drake::math::rpy2rotmat(rpy), xyz, 0, 0, 0, 1;
   }
 
   auto r3b1_and_r4b1_weld = std::allocate_shared<RigidBodyFrame>(
@@ -163,8 +166,8 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
       tree->FindBody("body1", "robot2"), T_r3_and_r4_to_r2);
 
   tree->AddFloatingJoint(DrakeJoint::QUATERNION,
-                        {r3b1->body_index, r4b1->body_index},
-                        r3b1_and_r4b1_weld);
+                         {r3b1->body_index, r4b1->body_index},
+                         r3b1_and_r4b1_weld);
 
   EXPECT_TRUE(tree->FindBody("body1", "robot1")
                   ->getJoint()
@@ -191,8 +194,9 @@ TEST_F(RigidBodyTreeTest, TestAddFloatingJointWeldToLink) {
 // with vector block input parameters. For more information, see:
 // https://github.com/RobotLocomotion/drake/issues/2634.
 TEST_F(RigidBodyTreeTest, TestDoKinematicsWithVectorBlocks) {
-  std::string file_name = Drake::getDrakePath() +
-          "/systems/plants/test/rigid_body_tree/two_dof_robot.urdf";
+  std::string file_name =
+      Drake::getDrakePath() +
+      "/systems/plants/test/rigid_body_tree/two_dof_robot.urdf";
   drake::parsers::urdf::AddRobotFromURDF(file_name, tree.get());
 
   VectorX<double> q;
@@ -202,10 +206,8 @@ TEST_F(RigidBodyTreeTest, TestDoKinematicsWithVectorBlocks) {
   q.setZero();
   v.setZero();
 
-  Eigen::VectorBlock<VectorX<double>>
-    q_block = q.head(q.size());
-  Eigen::VectorBlock<VectorX<double>>
-    v_block = v.head(v.size());
+  Eigen::VectorBlock<VectorX<double>> q_block = q.head(q.size());
+  Eigen::VectorBlock<VectorX<double>> v_block = v.head(v.size());
 
   KinematicsCache<double> cache = tree->doKinematics(q_block, v_block);
   EXPECT_TRUE(cache.hasV());
