+++ drake/systems/plants/test/testMagnetometer.cpp
@@ -1,5 +1,6 @@
 #include "gtest/gtest.h"
 
+#include "drake/math/roll_pitch_yaw.h"
 #include "drake/Path.h"
 #include "drake/systems/plants/RigidBodySystem.h"
 #include "drake/util/eigen_matrix_compare.h"
@@ -25,7 +26,7 @@ Vector3d getMagnetometerOutput(shared_ptr<RigidBodySystem> const& sys,
   VectorXd x0 = VectorXd::Zero(sys->getNumStates());
   auto const& tree = sys->getRigidBodyTree();
   x0.head(tree->number_of_positions()) = tree->getZeroConfiguration();
-  x0.segment(3, 4) = rpy2quat(rpy);
+  x0.segment(3, 4) = drake::math::rpy2quat(rpy);
   auto const& system_output = sys->output(0, x0, Vector4d::Zero());
   return system_output.tail<3>();
 }
