+++ drake/util/drakeGeometryUtil.h
@@ -76,278 +76,7 @@ void normalizeVec(
   }
 }
 
-/*
- * quat2x
- */
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 4, 1> quat2axis(
-    const Eigen::MatrixBase<Derived>& q) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
-  auto q_normalized = q.normalized();
-  auto s = std::sqrt(1.0 - q_normalized(0) * q_normalized(0)) +
-           std::numeric_limits<typename Derived::Scalar>::epsilon();
-  Eigen::Matrix<typename Derived::Scalar, 4, 1> a;
-
-  a << q_normalized.template tail<3>() / s, 2.0 * std::acos(q_normalized(0));
-  return a;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 3, 3> quat2rotmat(
-    const Eigen::MatrixBase<Derived>& q) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
-  auto q_normalized = q.normalized();
-  auto w = q_normalized(0);
-  auto x = q_normalized(1);
-  auto y = q_normalized(2);
-  auto z = q_normalized(3);
-
-  Eigen::Matrix<typename Derived::Scalar, 3, 3> M;
-  M.row(0) << w * w + x * x - y * y - z * z, 2.0 * x * y - 2.0 * w * z,
-      2.0 * x * z + 2.0 * w * y;
-  M.row(1) << 2.0 * x * y + 2.0 * w * z, w * w + y * y - x * x - z * z,
-      2.0 * y * z - 2.0 * w * x;
-  M.row(2) << 2.0 * x * z - 2.0 * w * y, 2.0 * y * z + 2.0 * w * x,
-      w * w + z * z - x * x - y * y;
-
-  return M;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 3, 1> quat2rpy(
-    const Eigen::MatrixBase<Derived>& q) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
-  auto q_normalized = q.normalized();
-  auto w = q_normalized(0);
-  auto x = q_normalized(1);
-  auto y = q_normalized(2);
-  auto z = q_normalized(3);
-
-  Eigen::Matrix<typename Derived::Scalar, 3, 1> ret;
-  ret << std::atan2(2.0 * (w * x + y * z), w * w + z * z - (x * x + y * y)),
-      std::asin(2.0 * (w * y - z * x)),
-      std::atan2(2.0 * (w * z + x * y), w * w + x * x - (y * y + z * z));
-  return ret;
-}
-
-template <typename Derived>
-Eigen::Quaternion<typename Derived::Scalar> quat2eigenQuaternion(
-    const Eigen::MatrixBase<Derived>& q) {
-  // The Eigen Quaterniond constructor when used with 4 arguments, uses the (w,
-  // x, y, z) ordering, just as we do.
-  // HOWEVER: when the constructor is called on a 4-element Vector, the elements
-  // must be in (x, y, z, w) order.
-  // So, the following two calls will give you the SAME quaternion:
-  // Quaternion<double>(q(0), q(1), q(2), q(3));
-  // Quaternion<double>(Vector4d(q(3), q(0), q(1), q(2)))
-  // which is gross and will cause you much pain.
-  // see:
-  // http://eigen.tuxfamily.org/dox/classEigen_1_1Quaternion.html#a91b6ea2cac13ab2d33b6e74818ee1490
-  //
-  // This method takes a nice, normal (w, x, y, z) order vector and gives you
-  // the Quaternion you expect.
-  return Eigen::Quaternion<typename Derived::Scalar>(q(0), q(1), q(2), q(3));
-}
-
-/*
- * axis2x
- */
-template <typename Derived>
-Eigen::Vector4d axis2quat(const Eigen::MatrixBase<Derived>& a) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
-  auto axis = a.template head<3>();
-  auto angle = a(3);
-  auto arg = 0.5 * angle;
-  auto c = std::cos(arg);
-  auto s = std::sin(arg);
-  Eigen::Vector4d ret;
-  ret << c, s * axis;
-  return ret;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 3, 3> axis2rotmat(
-    const Eigen::MatrixBase<Derived>& a) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
-  const auto& axis = (a.template head<3>()) / (a.template head<3>()).norm();
-  const auto& theta = a(3);
-  auto x = axis(0);
-  auto y = axis(1);
-  auto z = axis(2);
-  auto ctheta = std::cos(theta);
-  auto stheta = std::sin(theta);
-  auto c = 1 - ctheta;
-  Eigen::Matrix<typename Derived::Scalar, 3, 3> R;
-  R << ctheta + x * x * c, x * y * c - z * stheta, x * z * c + y * stheta,
-      y * x * c + z * stheta, ctheta + y * y * c, y * z * c - x * stheta,
-      z * x * c - y * stheta, z * y * c + x * stheta, ctheta + z * z * c;
-
-  return R;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 3, 1> axis2rpy(
-    const Eigen::MatrixBase<Derived>& a) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 4);
-  return quat2rpy(axis2quat(a));
-}
-
-/*
- * rotmat2x
- */
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 4, 1> rotmat2axis(
-    const Eigen::MatrixBase<Derived>& R) {
-  EIGEN_STATIC_ASSERT_MATRIX_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 3, 3);
-
-  typename Derived::Scalar theta = std::acos((R.trace() - 1.0) / 2.0);
-  Eigen::Vector4d a;
-  if (theta > std::numeric_limits<typename Derived::Scalar>::epsilon()) {
-    a << R(2, 1) - R(1, 2), R(0, 2) - R(2, 0), R(1, 0) - R(0, 1), theta;
-    a.head<3>() *= 1.0 / (2.0 * std::sin(theta));
-  } else {
-    a << 1.0, 0.0, 0.0, 0.0;
-  }
-  return a;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 4, 1> rotmat2quat(
-    const Eigen::MatrixBase<Derived>& M) {
-  EIGEN_STATIC_ASSERT_MATRIX_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 3, 3);
-  using namespace std;
-
-  typedef typename Derived::Scalar Scalar;
-  Eigen::Matrix<Scalar, 4, 3> A;
-  A.row(0) << 1.0, 1.0, 1.0;
-  A.row(1) << 1.0, -1.0, -1.0;
-  A.row(2) << -1.0, 1.0, -1.0;
-  A.row(3) << -1.0, -1.0, 1.0;
-  Eigen::Matrix<Scalar, 4, 1> B = A * M.diagonal();
-  Eigen::Index ind, max_col;
-  Scalar val = B.maxCoeff(&ind, &max_col);
-
-  Scalar w, x, y, z;
-  switch (ind) {
-    case 0: {
-      // val = trace(M)
-      w = sqrt(1.0 + val) / 2.0;
-      Scalar w4 = w * 4.0;
-      x = (M(2, 1) - M(1, 2)) / w4;
-      y = (M(0, 2) - M(2, 0)) / w4;
-      z = (M(1, 0) - M(0, 1)) / w4;
-      break;
-    }
-    case 1: {
-      // val = M(1, 1) - M(2, 2) - M(3, 3)
-      Scalar s = 2.0 * sqrt(1.0 + val);
-      w = (M(2, 1) - M(1, 2)) / s;
-      x = 0.25 * s;
-      y = (M(0, 1) + M(1, 0)) / s;
-      z = (M(0, 2) + M(2, 0)) / s;
-      break;
-    }
-    case 2: {
-      //  % val = M(2, 2) - M(1, 1) - M(3, 3)
-      Scalar s = 2.0 * (sqrt(1.0 + val));
-      w = (M(0, 2) - M(2, 0)) / s;
-      x = (M(0, 1) + M(1, 0)) / s;
-      y = 0.25 * s;
-      z = (M(1, 2) + M(2, 1)) / s;
-      break;
-    }
-    default: {
-      // val = M(3, 3) - M(2, 2) - M(1, 1)
-      Scalar s = 2.0 * (sqrt(1.0 + val));
-      w = (M(1, 0) - M(0, 1)) / s;
-      x = (M(0, 2) + M(2, 0)) / s;
-      y = (M(1, 2) + M(2, 1)) / s;
-      z = 0.25 * s;
-      break;
-    }
-  }
-
-  Eigen::Matrix<Scalar, 4, 1> q;
-  q << w, x, y, z;
-  return q;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 3, 1> rotmat2rpy(
-    const Eigen::MatrixBase<Derived>& R) {
-  EIGEN_STATIC_ASSERT_MATRIX_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 3, 3);
-  using namespace std;
-
-  Eigen::Matrix<typename Derived::Scalar, 3, 1> rpy;
-  rpy << atan2(R(2, 1), R(2, 2)),
-      atan2(-R(2, 0), sqrt(pow(R(2, 1), 2.0) + pow(R(2, 2), 2.0))),
-      atan2(R(1, 0), R(0, 0));
-  return rpy;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, Eigen::Dynamic, 1>
-rotmat2Representation(const Eigen::MatrixBase<Derived>& R, int rotation_type) {
-  typedef typename Derived::Scalar Scalar;
-  switch (rotation_type) {
-    case 0:
-      return Eigen::Matrix<Scalar, Eigen::Dynamic, 1>(0, 1);
-    case 1:
-      return rotmat2rpy(R);
-    case 2:
-      return rotmat2quat(R);
-    default:
-      throw std::runtime_error("rotation representation type not recognized");
-  }
-}
-
 DRAKEGEOMETRYUTIL_EXPORT int rotationRepresentationSize(int rotation_type);
-
-/*
- * rpy2x
- */
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 4, 1> rpy2quat(
-    const Eigen::MatrixBase<Derived>& rpy) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 3);
-  auto rpy_2 = (rpy / 2.0).array();
-  auto s = rpy_2.sin();
-  auto c = rpy_2.cos();
-
-  Eigen::Vector4d q;
-  q << c(0) * c(1) * c(2) + s(0) * s(1) * s(2),
-      s(0) * c(1) * c(2) - c(0) * s(1) * s(2),
-      c(0) * s(1) * c(2) + s(0) * c(1) * s(2),
-      c(0) * c(1) * s(2) - s(0) * s(1) * c(2);
-
-  q /= q.norm() + std::numeric_limits<typename Derived::Scalar>::epsilon();
-  return q;
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 4, 1> rpy2axis(
-    const Eigen::MatrixBase<Derived>& rpy) {
-  return quat2axis(rpy2quat(rpy));
-}
-
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 3, 3> rpy2rotmat(
-    const Eigen::MatrixBase<Derived>& rpy) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 3);
-  auto rpy_array = rpy.array();
-  auto s = rpy_array.sin();
-  auto c = rpy_array.cos();
-
-  Eigen::Matrix<typename Derived::Scalar, 3, 3> R;
-  R.row(0) << c(2) * c(1), c(2) * s(1) * s(0) - s(2) * c(0),
-      c(2) * s(1) * c(0) + s(2) * s(0);
-  R.row(1) << s(2) * c(1), s(2) * s(1) * s(0) + c(2) * c(0),
-      s(2) * s(1) * c(0) - c(2) * s(0);
-  R.row(2) << -s(1), c(1) * s(0), c(1) * c(0);
-
-  return R;
-}
 
 /*
  * rotation conversion gradient functions
@@ -542,35 +271,6 @@ drotmat2quat(const Eigen::MatrixBase<DerivedR>& R,
   return dq;
 }
 
-template <typename Derived>
-Eigen::Matrix<typename Derived::Scalar, 9, 3> drpy2rotmat(
-    const Eigen::MatrixBase<Derived>& rpy) {
-  EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(Eigen::MatrixBase<Derived>, 3);
-  auto rpy_array = rpy.array();
-  auto s = rpy_array.sin();
-  auto c = rpy_array.cos();
-
-  Eigen::Matrix<typename Derived::Scalar, 9, 3> dR;
-  dR.row(0) << 0, c(2) * -s(1), c(1) * -s(2);
-  dR.row(1) << 0, -s(1) * s(2), c(2) * c(1);
-  dR.row(2) << 0, -c(1), 0;
-  dR.row(3) << c(2) * s(1) * c(0) - s(2) * -s(0), c(2) * c(1) * s(0),
-      -s(2) * s(1) * s(0) - c(2) * c(0);
-  dR.row(4) << s(2) * s(1) * c(0) + c(2) * -s(0), s(2) * c(1) * s(0),
-      c(2) * s(1) * s(0) - s(2) * c(0);
-  dR.row(5) << c(1) * c(0), -s(1) * s(0), 0;
-  dR.row(6) << c(2) * s(1) * -s(0) + s(2) * c(0), c(2) * c(1) * c(0),
-      -s(2) * s(1) * c(0) + c(2) * s(0);
-  dR.row(7) << s(2) * s(1) * -s(0) - c(2) * c(0), s(2) * c(1) * c(0),
-      c(2) * s(1) * c(0) + s(2) * s(0);
-  dR.row(8) << c(1) * -s(0), -s(1) * c(0), 0;
-
-  return dR;
-}
-
-DRAKEGEOMETRYUTIL_EXPORT Eigen::Matrix3d rotz(double theta);
-DRAKEGEOMETRYUTIL_EXPORT void rotz(double theta, Eigen::Matrix3d& M,
-                                   Eigen::Matrix3d& dM, Eigen::Matrix3d& ddM);
 /*
  * cross product related
  */
@@ -745,167 +445,6 @@ void rpydot2angularvel(
   omega = E * rpydot;
 }
 
-template <typename Scalar>
-void cylindrical2cartesian(const Eigen::Matrix<Scalar, 3, 1>& m_cylinder_axis,
-                           const Eigen::Matrix<Scalar, 3, 1>& m_cylinder_x_dir,
-                           const Eigen::Matrix<Scalar, 3, 1>& cylinder_origin,
-                           const Eigen::Matrix<Scalar, 6, 1>& x_cylinder,
-                           const Eigen::Matrix<Scalar, 6, 1>& v_cylinder,
-                           Eigen::Matrix<Scalar, 6, 1>& x_cartesian,
-                           Eigen::Matrix<Scalar, 6, 1>& v_cartesian,
-                           Eigen::Matrix<Scalar, 6, 6>& J,
-                           Eigen::Matrix<Scalar, 6, 1>& Jdotv) {
-  Eigen::Matrix<Scalar, 3, 1> cylinder_axis =
-      m_cylinder_axis / m_cylinder_axis.norm();
-  Eigen::Matrix<Scalar, 3, 1> cylinder_x_dir =
-      m_cylinder_x_dir / m_cylinder_x_dir.norm();
-  Eigen::Matrix<Scalar, 3, 3> R_cylinder2cartesian;
-  R_cylinder2cartesian.col(0) = cylinder_x_dir;
-  R_cylinder2cartesian.col(1) = cylinder_axis.cross(cylinder_x_dir);
-  R_cylinder2cartesian.col(2) = cylinder_axis;
-  double radius = x_cylinder(0);
-  double theta = x_cylinder(1);
-  double c_theta = cos(theta);
-  double s_theta = sin(theta);
-  double height = x_cylinder(2);
-  double radius_dot = v_cylinder(0);
-  double theta_dot = v_cylinder(1);
-  double height_dot = v_cylinder(2);
-  Eigen::Matrix<Scalar, 3, 1> x_pos_cartesian;
-  x_pos_cartesian << radius * c_theta, radius * s_theta, height;
-  x_pos_cartesian = R_cylinder2cartesian * x_pos_cartesian + cylinder_origin;
-  Eigen::Matrix<Scalar, 3, 1> v_pos_cartesian;
-  v_pos_cartesian << radius * -s_theta * theta_dot + radius_dot * c_theta,
-      radius * c_theta * theta_dot + radius_dot * s_theta, height_dot;
-  v_pos_cartesian = R_cylinder2cartesian * v_pos_cartesian;
-  Eigen::Vector3d x_rpy_cylinder = x_cylinder.block(3, 0, 3, 1);
-  Eigen::Matrix<Scalar, 3, 3> R_tangent = rpy2rotmat(x_rpy_cylinder);
-  Eigen::Matrix<Scalar, 3, 3> R_tangent2cylinder;
-  Eigen::Matrix<Scalar, 3, 3> dR_tangent2cylinder;
-  Eigen::Matrix<Scalar, 3, 3> ddR_tangent2cylinder;
-  rotz(theta - M_PI / 2, R_tangent2cylinder, dR_tangent2cylinder,
-       ddR_tangent2cylinder);
-  Eigen::Matrix<Scalar, 3, 3> dR_tangent2cylinder_dtheta = dR_tangent2cylinder;
-  Eigen::Matrix<Scalar, 3, 3> R_cylinder = R_tangent2cylinder * R_tangent;
-  Eigen::Matrix<Scalar, 3, 3> R_cartesian = R_cylinder2cartesian * R_cylinder;
-  Eigen::Matrix<Scalar, 3, 1> x_rpy_cartesian = rotmat2rpy(R_cartesian);
-  x_cartesian.block(0, 0, 3, 1) = x_pos_cartesian;
-  x_cartesian.block(3, 0, 3, 1) = x_rpy_cartesian;
-  v_cartesian.block(0, 0, 3, 1) = v_pos_cartesian;
-  v_cartesian.block(3, 0, 3, 1) =
-      theta_dot * R_cylinder2cartesian.col(2) +
-      R_cylinder2cartesian * R_tangent2cylinder * v_cylinder.block(3, 0, 3, 1);
-  J = Eigen::Matrix<Scalar, 6, 6>::Zero();
-  J.block(0, 0, 3, 1) << c_theta, s_theta, 0;
-  J.block(0, 1, 3, 1) << radius * -s_theta, radius * c_theta, 0;
-  J.block(0, 2, 3, 1) << 0, 0, 1;
-  J.block(0, 0, 3, 3) = R_cylinder2cartesian * J.block(0, 0, 3, 3);
-  J.block(3, 1, 3, 1) = R_cylinder2cartesian.col(2);
-  J.block(3, 3, 3, 3) = R_cylinder2cartesian * R_tangent2cylinder;
-  Eigen::Matrix<Scalar, 3, 3> dJ1_dradius = Eigen::Matrix<Scalar, 3, 3>::Zero();
-  dJ1_dradius(0, 1) = -s_theta;
-  dJ1_dradius(1, 1) = c_theta;
-  Eigen::Matrix<Scalar, 3, 3> dJ1_dtheta = Eigen::Matrix<Scalar, 3, 3>::Zero();
-  dJ1_dtheta(0, 0) = -s_theta;
-  dJ1_dtheta(0, 1) = -radius * c_theta;
-  dJ1_dtheta(1, 0) = c_theta;
-  dJ1_dtheta(1, 1) = -radius * s_theta;
-  Jdotv.block(0, 0, 3, 1) = R_cylinder2cartesian * (dJ1_dradius * radius_dot +
-                                                    dJ1_dtheta * theta_dot) *
-                            v_cylinder.block(0, 0, 3, 1);
-  Jdotv.block(3, 0, 3, 1) = R_cylinder2cartesian * dR_tangent2cylinder_dtheta *
-                            theta_dot * v_cylinder.block(3, 0, 3, 1);
-}
-
-template <typename Scalar>
-void cartesian2cylindrical(const Eigen::Matrix<Scalar, 3, 1>& m_cylinder_axis,
-                           const Eigen::Matrix<Scalar, 3, 1>& m_cylinder_x_dir,
-                           const Eigen::Matrix<Scalar, 3, 1>& cylinder_origin,
-                           const Eigen::Matrix<Scalar, 6, 1>& x_cartesian,
-                           const Eigen::Matrix<Scalar, 6, 1>& v_cartesian,
-                           Eigen::Matrix<Scalar, 6, 1>& x_cylinder,
-                           Eigen::Matrix<Scalar, 6, 1>& v_cylinder,
-                           Eigen::Matrix<Scalar, 6, 6>& J,
-                           Eigen::Matrix<Scalar, 6, 1>& Jdotv) {
-  Eigen::Matrix<Scalar, 3, 1> cylinder_axis =
-      m_cylinder_axis / m_cylinder_axis.norm();
-  Eigen::Matrix<Scalar, 3, 1> cylinder_x_dir =
-      m_cylinder_x_dir / m_cylinder_x_dir.norm();
-  Eigen::Matrix<Scalar, 3, 3> R_cylinder2cartesian;
-  R_cylinder2cartesian.col(0) = cylinder_x_dir;
-  R_cylinder2cartesian.col(1) = cylinder_axis.cross(cylinder_x_dir);
-  R_cylinder2cartesian.col(2) = cylinder_axis;
-  Eigen::Matrix<Scalar, 3, 3> R_cartesian2cylinder =
-      R_cylinder2cartesian.transpose();
-  Eigen::Matrix<Scalar, 3, 1> x_pos_cylinder =
-      R_cartesian2cylinder * (x_cartesian.block(0, 0, 3, 1) - cylinder_origin);
-  Eigen::Matrix<Scalar, 3, 1> v_pos_cylinder =
-      R_cartesian2cylinder * v_cartesian.block(0, 0, 3, 1);
-  double radius = sqrt(pow(x_pos_cylinder(0), 2) + pow(x_pos_cylinder(1), 2));
-  double radius_dot = (x_pos_cylinder(0) * v_pos_cylinder(0) +
-                       x_pos_cylinder(1) * v_pos_cylinder(1)) /
-                      radius;
-  double theta = atan2(x_pos_cylinder(1), x_pos_cylinder(0));
-  double radius_square = pow(radius, 2);
-  double radius_cubic = pow(radius, 3);
-  double radius_quad = pow(radius, 4);
-  double theta_dot = (-x_pos_cylinder(1) * v_pos_cylinder(0) +
-                      x_pos_cylinder(0) * v_pos_cylinder(1)) /
-                     radius_square;
-  double height = x_pos_cylinder(2);
-  double height_dot = v_pos_cylinder(2);
-  x_cylinder(0) = radius;
-  x_cylinder(1) = theta;
-  x_cylinder(2) = height;
-  v_cylinder(0) = radius_dot;
-  v_cylinder(1) = theta_dot;
-  v_cylinder(2) = height_dot;
-  Eigen::Matrix<Scalar, 3, 3> R_tangent2cylinder;
-  Eigen::Matrix<Scalar, 3, 3> dR_tangent2cylinder;
-  Eigen::Matrix<Scalar, 3, 3> ddR_tangent2cylinder;
-  rotz(theta - M_PI / 2, R_tangent2cylinder, dR_tangent2cylinder,
-       ddR_tangent2cylinder);
-  Eigen::Matrix<Scalar, 3, 3> R_cylinder2tangent =
-      R_tangent2cylinder.transpose();
-  Eigen::Vector3d x_rpy_cartesian = x_cartesian.block(3, 0, 3, 1);
-  Eigen::Matrix<Scalar, 3, 3> R_cartesian = rpy2rotmat(x_rpy_cartesian);
-  x_cylinder.block(3, 0, 3, 1) =
-      rotmat2rpy(R_cylinder2tangent * R_cartesian2cylinder * R_cartesian);
-  J = Eigen::Matrix<Scalar, 6, 6>::Zero();
-  Eigen::Matrix<Scalar, 6, 6> Jdot = Eigen::Matrix<Scalar, 6, 6>::Zero();
-  J(0, 0) = x_pos_cylinder(0) / radius;
-  J(0, 1) = x_pos_cylinder(1) / radius;
-  J(1, 0) = -x_pos_cylinder(1) / radius_square;
-  J(1, 1) = x_pos_cylinder(0) / radius_square;
-  J(2, 2) = 1.0;
-  J.block(0, 0, 3, 3) = J.block(0, 0, 3, 3) * R_cartesian2cylinder;
-  Jdot(0, 0) =
-      pow(x_pos_cylinder(1), 2) / radius_cubic * v_pos_cylinder(0) -
-      x_pos_cylinder(0) * x_pos_cylinder(1) / radius_cubic * v_pos_cylinder(1);
-  Jdot(0, 1) = -x_pos_cylinder(0) * x_pos_cylinder(1) / radius_cubic *
-                   v_pos_cylinder(0) +
-               pow(x_pos_cylinder(0), 2) / radius_cubic * v_pos_cylinder(1);
-  Jdot(1, 0) = 2 * x_pos_cylinder(0) * x_pos_cylinder(1) / radius_quad *
-                   v_pos_cylinder(0) +
-               (pow(x_pos_cylinder(1), 2) - pow(x_pos_cylinder(0), 2)) /
-                   radius_quad * v_pos_cylinder(1);
-  Jdot(1, 1) = (pow(x_pos_cylinder(1), 2) - pow(x_pos_cylinder(0), 2)) /
-                   radius_quad * v_pos_cylinder(0) -
-               2 * x_pos_cylinder(0) * x_pos_cylinder(1) / radius_quad *
-                   v_pos_cylinder(1);
-  Jdot.block(0, 0, 3, 3) = Jdot.block(0, 0, 3, 3) * R_cartesian2cylinder;
-  v_cylinder.block(3, 0, 3, 1) = R_cylinder2tangent * R_cartesian2cylinder *
-                                     v_cartesian.block(3, 0, 3, 1) -
-                                 theta_dot * R_cylinder2tangent.col(2);
-  J.block(3, 0, 3, 3) = R_cylinder2tangent.col(2) * -J.block(1, 0, 1, 3);
-  J.block(3, 3, 3, 3) = R_cylinder2tangent * R_cartesian2cylinder;
-  Jdot.block(3, 0, 3, 3) = dR_tangent2cylinder.row(2).transpose() *
-                               -J.block(1, 0, 1, 3) * theta_dot +
-                           R_cylinder2tangent.col(2) * -Jdot.block(1, 0, 1, 3);
-  Jdot.block(3, 3, 3, 3) =
-      dR_tangent2cylinder.transpose() * theta_dot * R_cartesian2cylinder;
-  Jdotv = Jdot * v_cartesian;
-}
 /*
  * spatial transform functions
  */
